+++ definitions/data_bag_merge.rb
@@ -6,17 +6,17 @@
 #
 # Licensed under the MIT license
 
-define :data_bag_merge, :format => 'default' do
+define :data_bag_merge, :bag_format => 'default' do
   if params[:data_bag].nil?
     bag_item_name = params[:name].split('/')
     params[:data_bag] = bag_item_name[0]
     params[:item] = bag_item_name[1]
   end
 
-  data_bag_data = Chef::DataBagItem.load(params[:data_bag], params[:item]).to_hash
+  data_bag_data = Chef::DataBagItem.load(params[:data_bag], params[:item]).raw_data
   data_bag_data.delete 'id'
 
-  case params[:format]
+  case params[:bag_format]
   when 'default'
     Chef::Mixin::DeepMerge.deep_merge! data_bag_data, node.default
   when 'override'