+++ recipes/environment.rb
@@ -12,13 +12,13 @@ config = node['data-bag-merge']['environments']
 if config['encrypted']
   encrypted_data_bag_merge "#{config['data_bag_name']}/#{item_name}" do
     secret_path config['secret_path']
-    format config['format']
+    bag_format config['bag_format']
     data_bag config['data_bag_name']
     item item_name
   end
 else
   data_bag_merge "#{config['data_bag_name']}/#{item_name}" do
-    format config['format']
+    bag_format config['bag_format']
     data_bag config['data_bag_name']
     item item_name
   end